refactor: create paymentService
